+++ internal/retry/retry.go
@@ -1,4 +1,4 @@
-package http
+package retry
 
 import (
 	"fmt"
